Remove AVF_INFO content
Sony uses an obscure (proprietary?) indexing system on the SD card.
Each time you empty `DCIM` folder (or folder containing video file),
a message appears on the camera indicating an error and you got
errors in playback (because it tries to access files which have been
removed).

So let's reset this index by removing these files.
The camera still indicates there is an error on indexing, but you
just have to click `OK` and the camera resets the index.
As a consequence, you don't have errors in playback anymore.